[nas-assets] More testing related to S3 use

+++ app/models/traits/HasImage.php
@@ -47,12 +47,17 @@ trait HasImage
             $public_path = 'uploads/img/'.$this->collection.'/'.$this->image;
             $full_path = app_path().'/../public/'.$public_path;
 
-            $s3 = new \Laramongo\Nas\S3;
-            $result = $s3->sendFile($public_path);    
-
-            if($result)
+            $s3 = app()->s3;
+            if( $s3 )
             {
-                unlink($full_path);
+                $result = $s3->sendFile($public_path);    
+
+                if($result)
+                {
+                    unlink($full_path);
+                }
+
+                return $result;
             }
         }
